relocate test
