+++ composer/distributed/__init__.py
@@ -3,6 +3,7 @@
 
 """Distributed training."""
 
+# TODO(boweny) include fsdp2 after we deprecate torch-cpu 2.5
 from composer.distributed.dist_strategy import (
     DDPSyncStrategy,
     ddp_sync_context,
